Rename to texture-renderer.ts
